Fix warnings when no progress files are present at startup

+++ inst/app/server.R
@@ -48,8 +48,12 @@ finalStates = c("8.1", "9.1", "8.3", "7.3", "8.4", "9.2", "9.3")
 mcmcProgressFolder <- "mcmc-progress/progress/"
 mcmcPidFolder <- "mcmc-progress/pid/"
 
-file.remove(paste(mcmcProgressFolder, list.files(path=mcmcProgressFolder), sep=""))
-file.remove(paste(mcmcPidFolder, list.files(path=mcmcPidFolder), sep=""))
+if (length(list.files(path=mcmcProgressFolder)) > 0) {
+  file.remove(paste(mcmcProgressFolder, list.files(path=mcmcProgressFolder), sep=""))
+}
+if (length(list.files(path=mcmcPidFolder)) > 0) {
+  file.remove(paste(mcmcPidFolder, list.files(path=mcmcPidFolder), sep=""))
+}
 
 shinyServer(function(input, output, session) {
   # DEBUG
